+++ src/Chainweb/Chainweb.hs
@@ -331,7 +331,7 @@ validatingMempoolConfig cid v gl gp mv = Mempool.InMemConfig
 
 data StartedChainweb logger where
   StartedChainweb :: (CanReadablePayloadCas cas, Logger logger) => !(Chainweb logger cas) -> StartedChainweb logger
-  Replayed :: !Cut -> !Cut -> StartedChainweb logger
+  Replayed :: !Cut -> !(Maybe Cut) -> StartedChainweb logger
 
 data ChainwebStatus
     = ProcessStarted
@@ -369,7 +369,8 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
     -- Garbage Collection
     -- performed before PayloadDb and BlockHeaderDb used by other components
     logFunctionJson logger Info PruningDatabases
-    logg Info "start pruning databases"
+    when (_cutPruneChainDatabase (_configCuts conf) /= GcNone) $
+        logg Info "start pruning databases"
     case _cutPruneChainDatabase (_configCuts conf) of
         GcNone -> return ()
         GcHeaders ->
@@ -378,39 +379,48 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
             pruneAllChains (pruningLogger "headers-checked") rocksDb v [CheckPayloads, CheckFull]
         GcFull ->
             fullGc (pruningLogger "full") rocksDb v
-    logg Info "finished pruning databases"
+    when (_cutPruneChainDatabase (_configCuts conf) /= GcNone) $
+        logg Info "finished pruning databases"
     logFunctionJson logger Info InitializingChainResources
 
-    logg Info "start initializing chain resources"
-    concurrentWith
-        -- initialize chains concurrently
-        (\cid x -> do
-            let mcfg = validatingMempoolConfig cid v (_configBlockGasLimit conf) (_configMinGasPrice conf)
-            -- NOTE: the gas limit may be set based on block height in future, so this approach may not be valid.
-            let maxGasLimit = fromIntegral <$> maxBlockGasLimit v maxBound
-            case maxGasLimit of
-                Just maxGasLimit'
-                    | _configBlockGasLimit conf > maxGasLimit' ->
-                        logg Warn $ T.unwords
-                            [ "configured block gas limit is greater than the"
-                            , "maximum for this chain; the maximum will be used instead"
-                            ]
-                _ -> return ()
-            withChainResources
-                v
-                cid
-                rocksDb
-                (chainLogger cid)
-                mcfg
-                payloadDb
-                pactDbDir
-                (pactConfig maxGasLimit)
-                x
-        )
+    txFailuresCounter <- newCounter @"txFailures"
+    let monitorTxFailuresCounter =
+            runForever (logFunctionText logger) "monitor txFailuresCounter" $ do
+                approximateThreadDelay 60000000 {- 1 minute -}
+                logFunctionCounter logger Info . (:[]) =<<
+                    roll txFailuresCounter
+    logg Debug "start initializing chain resources"
+    withAsync monitorTxFailuresCounter $ \_ ->
+        concurrentWith
+            -- initialize chains concurrently
+            (\cid x -> do
+                let mcfg = validatingMempoolConfig cid v (_configBlockGasLimit conf) (_configMinGasPrice conf)
+                -- NOTE: the gas limit may be set based on block height in future, so this approach may not be valid.
+                let maxGasLimit = fromIntegral <$> maxBlockGasLimit v maxBound
+                case maxGasLimit of
+                    Just maxGasLimit'
+                        | _configBlockGasLimit conf > maxGasLimit' ->
+                            logg Warn $ T.unwords
+                                [ "configured block gas limit is greater than the"
+                                , "maximum for this chain; the maximum will be used instead"
+                                ]
+                    _ -> return ()
+                withChainResources
+                    v
+                    cid
+                    rocksDb
+                    (chainLogger cid)
+                    mcfg
+                    payloadDb
+                    pactDbDir
+                    (pactConfig maxGasLimit)
+                    txFailuresCounter
+                    x
+            )
 
         -- initialize global resources after all chain resources are initialized
         (\cs -> do
-            logg Info "finished initializing chain resources"
+            logg Debug "finished initializing chain resources"
             global (HM.fromList $ zip cidsList cs)
         )
         cidsList
@@ -464,11 +474,11 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
             !cutLogger = setComponent "cut" logger
             !mgr = _peerResManager peer
 
-        logg Info "start initializing cut resources"
+        logg Debug "start initializing cut resources"
         logFunctionJson logger Info InitializingCutResources
 
         withCutResources cutConfig peer cutLogger rocksDb webchain payloadDb mgr pact $ \cuts -> do
-            logg Info "finished initializing cut resources"
+            logg Debug "finished initializing cut resources"
 
             let !mLogger = setComponent "miner" logger
                 !mConf = _configMining conf
@@ -479,7 +489,7 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
             throttler <- mkGenericThrottler $ _throttlingRate throt
             putPeerThrottler <- mkPutPeerThrottler $ _throttlingPeerRate throt
             mempoolThrottler <- mkMempoolThrottler $ _throttlingMempoolRate throt
-            logg Info "initialized throttlers"
+            logg Debug "initialized throttlers"
 
             -- synchronize pact dbs with latest cut before we start the server
             -- and clients and begin mining.
@@ -503,25 +513,27 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
                     unsafeMkCut v <$> readHighestCutHeaders v (logFunctionText logger) webchain (cutHashesTable rocksDb)
                 lowerBoundCut <-
                     tryLimitCut webchain (fromMaybe 0 $ _cutInitialBlockHeightLimit $ _configCuts conf) highestCut
-                upperBoundCut <-
-                    tryLimitCut webchain (fromMaybe maxBound $ _cutFastForwardBlockHeightLimit $ _configCuts conf) highestCut
+                upperBoundCut <- forM (_cutFastForwardBlockHeightLimit $ _configCuts conf) $ \upperBound ->
+                    tryLimitCut webchain upperBound highestCut
                 let
-                    replayOneChain :: (ChainResources logger, (BlockHeader, BlockHeader)) -> IO ()
+                    replayOneChain :: (ChainResources logger, (BlockHeader, Maybe BlockHeader)) -> IO ()
                     replayOneChain (cr, (l, u)) = do
                         let chainPact = _chainResPact cr
                         let logCr = logFunctionText
                                 $ addLabel ("component", "pact")
                                 $ addLabel ("sub-component", "init")
                                 $ _chainResLogger cr
-                        logCr Info $ "pact db replaying between blocks "
-                            <> T.pack (show (_blockHeight l, _blockHash l)) <> " and "
-                            <> T.pack (show (_blockHeight u, _blockHash u))
                         void $ _pactReadOnlyReplay chainPact l u
                         logCr Info "pact db synchronized"
-                mapConcurrently_ replayOneChain $
-                    HM.intersectionWith (,)
-                        pactSyncChains
-                        (HM.intersectionWith (,) (_cutMap lowerBoundCut) (_cutMap upperBoundCut))
+                let bounds =
+                        HM.intersectionWith (,)
+                            pactSyncChains
+                            (HM.mapWithKey
+                                (\cid bh ->
+                                    (bh, (HM.! cid) . _cutMap <$> upperBoundCut))
+                                (_cutMap lowerBoundCut)
+                            )
+                mapConcurrently_ replayOneChain bounds
                 logg Info "finished fast forward replay"
                 logFunctionJson logger Info PactReplaySuccessful
                 inner $ Replayed lowerBoundCut upperBoundCut
@@ -539,7 +551,7 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
                 synchronizePactDb pactSyncChains newCut
                 logg Info "finished replaying Pact DBs to fast forward cut"
                 logFunctionJson logger Info PactReplaySuccessful
-                inner $ Replayed initialCut newCut
+                inner $ Replayed initialCut (Just newCut)
             else do
                 initialCut <- _cut mCutDb
                 logg Info "start synchronizing Pact DBs to initial cut"
@@ -547,7 +559,7 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
                 synchronizePactDb pactSyncChains initialCut
                 logg Info "finished synchronizing Pact DBs to initial cut"
                 withPactData cs cuts $ \pactData -> do
-                    logg Info "start initializing miner resources"
+                    logg Debug "start initializing miner resources"
                     logFunctionJson logger Info InitializingMinerResources
 
                     withMiningCoordination mLogger mConf mCutDb $ \mc ->
@@ -558,7 +570,7 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
                         --
                         withMinerResources mLogger (_miningInNode mConf) cs mCutDb mc $ \m -> do
                             logFunctionJson logger Info ChainwebStarted
-                            logg Info "finished initializing miner resources"
+                            logg Debug "finished initializing miner resources"
                             let !haddr = _peerConfigAddr $ _p2pConfigPeer $ _configP2p conf
                             inner $ StartedChainweb Chainweb
                                 { _chainwebHostAddress = haddr
@@ -627,12 +639,8 @@ withChainwebInternal conf logger peer serviceSock rocksDb pactDbDir backupDir re
                     $ addLabel ("component", "pact")
                     $ addLabel ("sub-component", "init")
                     $ _chainResLogger cr
-            let hsh = _blockHash bh
-            let h = _blockHeight bh
-            logCr Info $ "pact db synchronizing to block "
-                <> T.pack (show (h, hsh))
             void $ _pactSyncToBlock pact bh
-            logCr Info "pact db synchronized"
+            logCr Debug "pact db synchronized"
 
 -- -------------------------------------------------------------------------- --
 -- Throttling
@@ -692,7 +700,7 @@ runChainweb
     -> ((NowServing -> NowServing) -> IO ())
     -> IO ()
 runChainweb cw nowServing = do
-    logg Info "start chainweb node"
+    logg Debug "start chainweb node"
     mkValidationMiddleware <- interleaveIO $
         OpenAPIValidation.mkValidationMiddleware (_chainwebLogger cw) (_chainwebVersion cw) (_chainwebManager cw)
     p2pValidationMiddleware <-
@@ -774,6 +782,20 @@ runChainweb cw nowServing = do
     pactDbsToServe :: [(ChainId, PactServerData logger tbl)]
     pactDbsToServe = _chainwebPactData cw
 
+     loggServerError msg (Just r) e =
+         "HTTP server error (" <> msg <> "): " <> sshow e <> ". Request: " <> sshow r
+     loggServerError msg Nothing e =
+         "HTTP server error (" <> msg <> "): " <> sshow e
+
+     logWarpException msg clientClosedConnectionsCounter r e
+         | Just InsecureConnectionDenied <- fromException e =
+             return ()
+         | Just ClientClosedConnectionPrematurely <- fromException e =
+             inc clientClosedConnectionsCounter
+         | otherwise =
+             when (defaultShouldDisplayException e) $
+                logg Debug $ loggServerError msg r e
+
     -- P2P Server
 
     serverSettings :: Counter "clientClosedConnections" -> Settings
@@ -787,7 +809,7 @@ runChainweb cw nowServing = do
                     inc closedConnectionsCounter
                 | otherwise ->
                     when (defaultShouldDisplayException e) $
-                        logg Warn $ loggServerError r e
+                        logg Debug $ loggServerError r e
             )
         & setBeforeMainLoop (nowServing (nowServingP2PAPI .~ True))